+++ rts/operator/NestedLoopJoin.cpp
@@ -11,7 +11,7 @@
 // or send a letter to Creative Commons, 171 Second Street, Suite 300,
 // San Francisco, California, 94105, USA.
 //---------------------------------------------------------------------------
-NestedLoopJoin::NestedLoopJoin(Operator* left,Operator* right,unsigned expectedOutputCardinality)
+NestedLoopJoin::NestedLoopJoin(Operator* left,Operator* right,double expectedOutputCardinality)
    : Operator(expectedOutputCardinality),left(left),right(right)
    // Constructor
 {
