+++ rts/operator/Union.cpp
@@ -11,7 +11,7 @@
 // or send a letter to Creative Commons, 171 Second Street, Suite 300,
 // San Francisco, California, 94105, USA.
 //---------------------------------------------------------------------------
-Union::Union(const std::vector<Operator*>& parts,const std::vector<std::vector<Register*> >& mappings,const std::vector<std::vector<Register*> >& initializations,unsigned expectedOutputCardinality)
+Union::Union(const std::vector<Operator*>& parts,const std::vector<std::vector<Register*> >& mappings,const std::vector<std::vector<Register*> >& initializations,double expectedOutputCardinality)
    : Operator(expectedOutputCardinality),parts(parts),mappings(mappings),initializations(initializations)
    // Constructor
 {
